Extract fixed-float FpML swap mapping
