+++ docs/developer/fpml_import.rst
@@ -45,9 +45,11 @@ Internal Module Boundaries
 
 ``trellis.io.fpml.importer`` owns secure document inspection and
 ``trellis.io.fpml.contracts`` owns immutable body-free reports.
-``trellis.io.fpml.normalizer`` remains the stable normalization facade and,
-until the product-mapping extraction tickets close, contains the admitted
-swap, swaption, and cap/floor mappings. The internal
+``trellis.io.fpml.normalizer`` remains the stable normalization facade and
+contains the admitted swaption and cap/floor mappings until their extraction
+tickets close. The internal ``trellis.io.fpml._normalization_swap`` module owns
+fixed-float swap validation, semantic mapping, provenance, and the
+product-specific historical-fixing rejection. The internal
 ``trellis.io.fpml._normalization_common`` module owns product-neutral XML
 access, exact blocker and provenance construction, calendars, date and
 frequency conventions, regular schedule validation, and bounded stream
